Rename src/editors to src/sources
